refactor: extract snapshot-writer from container-runner
Moves snapshot writing functions (writeTasksSnapshot, writeGroupsSnapshot,
AvailableGroup) from container-runner.ts into a dedicated
snapshot-writer.ts module. Re-exports maintained in container-runner.ts
for backwards compatibility — no changes needed in existing importers.

Zero behavior change — all 352 tests pass.
